Add TUI, add IAM modules, add claude skills

+++ pkg/core/repl/identity.go
@@ -11,6 +11,9 @@ func (r *REPL) cmdIdentity(repl *REPL, args []string) error {
 
 	switch args[0] {
 	case "add":
+		if len(args) > 1 && args[1] == "help" {
+			return r.showIdentityAddHelp()
+		}
 		err := r.identityManager.AddIdentity(args[1:])
 		if err == nil {
 			// Update prompt in case identity was added and switched to
@@ -22,10 +25,27 @@ func (r *REPL) cmdIdentity(repl *REPL, args []string) error {
 	case "show":
 		return r.identityManager.ShowCurrent()
 	case "switch":
+		if len(args) > 1 && args[1] == "help" {
+			return r.showIdentitySwitchHelp()
+		}
 		if len(args) < 2 {
 			return NewInvalidArgumentsError("identity switch requires identity name")
 		}
 		err := r.identityManager.SwitchIdentity(args[1])
+		if err == nil {
+			r.UpdatePrompt()
+			r.showPmapperHint()
+		}
+		return err
+	case "check":
+		if len(args) > 1 && args[1] == "help" {
+			return r.showIdentityCheckHelp()
+		}
+		name := ""
+		if len(args) > 1 {
+			name = args[1]
+		}
+		err := r.identityManager.CheckAdmin(name)
 		if err == nil {
 			r.UpdatePrompt()
 		}
@@ -33,6 +53,9 @@ func (r *REPL) cmdIdentity(repl *REPL, args []string) error {
 	case "refresh":
 		return r.identityManager.RefreshCurrentIdentity()
 	case "clear", "remove":
+		if len(args) > 1 && args[1] == "help" {
+			return r.showIdentityClearHelp()
+		}
 		return r.identityManager.RemoveIdentity(args[1:])
 	case "help":
 		return r.showIdentityHelp()
